pyproject.toml
[tool.poetry]
name = "qstrain_cleaner"
version = "0.1.0"
description = ""
authors = ["cardionaut"]

[tool.poetry.dependencies]
python = "^3.9.13"
loguru = "^0.6.0"
hydra-core = "^1.3"
openpyxl = "^3.0.10"
matplotlib = "^3.5.3"
seaborn = "^0.11.2"
pylint = "^2.15.0"
jupyter = "^1.0.0"
notebook = "^6.4.12"
omegaconf = "^2.2"
xlsxwriter = "^3.0.7"

[tool.poetry.group.dev.dependencies]
autohooks = "^23.1.0"
autohooks-plugin-black = "^22.11.0"
autohooks-plugin-isort = "^22.8.0"
pytest = "^7.2.1"
black = "^22.6.0"
isort = "^5.10.1"

[tool.autohooks]
mode = "poetry"
pre-commit = ["autohooks.plugins.black", "autohooks.plugins.isort"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
skip-string-normalization = true

[tool.isort]
profile = "black"
